requirements.txt
alembic==1.11.1
amqp==5.1.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.99.0
fastapi-cache2==0.2.1
fastapi-users==12.0.0
fastapi-users-db-sqlalchemy==6.0.0
flower==2.0.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
humanize==4.7.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.1
packaging==23.1
passlib==1.7.4
pendulum==2.1.2
pluggy==1.2.0
prometheus-client==0.17.0
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==1.10.10
PyJWT==2.7.0
pytest==7.4.0
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
tomli==2.0.1
tornado==6.3.2
typing_extensions==4.7.0
tzdata==2023.3
ujson==5.8.0
uvicorn==0.22.0
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
